Implement step 2 and 3 of the decoding process
The majority of step 2 and 3 was implemented. This includes: sampling the image
to an array of bits while skipping the reference grid (if symbol is a full symbol),
reading the sampled symbol to a ordered list of bits, marking erasures (will be used
by Reed-Solomon error detection in the future), removing the stuffed bits, and
finally leading the prepared bitstring to the decoder.